Divide GameConfig into Regulation and DebugConfig
