+++ kodecks/src/profile.rs
@@ -1,26 +1,28 @@
-use crate::{config::GameConfig, player::PlayerConfig, regulation::Regulation, scenario::Scenario};
+use crate::{player::PlayerConfig, regulation::Regulation};
+use bitflags::bitflags;
 use serde::{Deserialize, Serialize};
-use std::fmt;
 
-#[derive(Serialize, Deserialize)]
+#[derive(Debug, Default, Clone, Serialize, Deserialize)]
 pub struct GameProfile {
-    pub config: GameConfig,
     pub regulation: Regulation,
+    pub debug: DebugConfig,
     pub players: Vec<PlayerConfig>,
     pub bots: Vec<BotConfig>,
-
-    #[serde(skip)]
-    pub scenario: Option<Box<dyn Scenario>>,
 }
 
-impl fmt::Debug for GameProfile {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("GameProfile")
-            .field("config", &self.config)
-            .field("regulation", &self.regulation)
-            .field("players", &self.players)
-            .field("bots", &self.bots)
-            .finish()
+#[derive(Debug, Default, Clone, Copy, Serialize, Deserialize)]
+pub struct DebugConfig {
+    pub rng_seed: Option<u64>,
+    pub no_deck_shuffle: bool,
+    pub no_player_shuffle: bool,
+    pub flags: DebugFlags,
+}
+
+bitflags! {
+    #[derive(Debug, Clone, Copy, Default, PartialEq, Eq, Serialize, Deserialize)]
+    pub struct DebugFlags: u8 {
+        const DEBUG_COMMAND = 0b00000001;
+        const IGNORE_COST = 0b00000010;
     }
 }
